requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
gunicorn==19.9.0
idna==2.8
idna-ssl==1.1.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
multidict==4.5.2
python-dotenv==0.10.3
requests==2.22.0
rollbar==0.14.7
six==1.12.0
slackclient==2.1.0
typing-extensions==3.7.4
urllib3==1.25.3
Werkzeug==0.15.5
yarl==1.3.0
